+++ icu4c/source/i18n/hextouni.cpp
@@ -305,7 +305,7 @@ void HexToUnicodeTransliterator::handleTransliterate(Replaceable& text, UTransPo
                         }
                         break;
                     }
-                    int8_t digit = u_digit(text.charAt(curs), 16);
+                    int32_t digit = u_digit(text.charAt(curs), 16);
                     if (digit < 0) {
                         break;
                     }
